requirements.txt
Pillow
pyyaml